Addition of README.txt
Since there was no README file yet in the repository, I added this file so that everyone can do this part of the assignment.
Please read the file carefully before making changes. If everyone works the same way, it will be easier to follow the progress of everyone.

Ben Van den Broeck